Fix PR closing race condition
Publication of https://github.com/openid/publication/pull/90 aborted

This appears to be because github marked the merge request as merged,
because the publication tool had merged the commits in the merge request.
Then the cleanup step tried to close the already closed merge request,
resulting in a failure that prevented the publication step running.

As github is closing the MR we don't need to do so manually, so remove
that step. This will mean the branch isn't deleted I think, but that seems
better than having an unreliable tool.